Add generic ContentTabs component
